nixos/systemd-boot: Add rememberLastChoice option.
The systemd-boot configuration lacks the ability to remember the
last used choice. This feature is included with systemd-boot via the
`default @saved` option, which saves the last choice into an EFI
variable. Having this option is useful for when users have mutliple OSes
installed on a single machine but the boot loader is managed by nix-os.

stuff