nixos/systemd-boot: Add current generation entry
To be able to automatically boot into the newest version when using
rememberLastChoice, we use a custom entry which always points to the
current generation. That way if you add a generation the current
generation entry will now point to the new generation, but systemd-
boot will still use our current generation entry and thus boot into
the newest version.